+++ xl2roefact/libutils.py
@@ -19,9 +19,11 @@ from fractions import Fraction
 
 # NOTE - ready, test PASS @231123
 def isnumber(a_string: str) -> bool:
-    """test is a string is valid as any kind of number
+    """ test is a string is valid as any kind of number
+
     Arguments:
-     - input string
+        - `a_string` input string
+
     Returns:
         - `True` if input string is valid as any kind of number, orherwise `False`
     """
@@ -39,17 +41,19 @@ def isnumber(a_string: str) -> bool:
 
 # NOTE - ready, test PASS @231123
 def find_str_in_list(list_of_str_to_find: list, list_to_search: list) -> int:
-    """find a substring in elements of list()
+    """ find a substring from `list_of_str_to_find` in elements of `list_to_search`
+
     Arguments:
         - `list_of_str_to_find` - list of strings to search for
-        - `list_to_search` - here search for `str_to_find`
+        - `list_to_search` - liste where to search for substrings
+
     Returns:
         - `index` - the index of list item which contains `str_to_find` (first found) or `None` if not found
     """
     __found = False
     for __crt_str_to_search_for in list_of_str_to_find:
         __tmp_list = [
-            x.find(__crt_str_to_search_for)
+            x.find(str(__crt_str_to_search_for))
             for i, x in enumerate(
                 [x.lower().replace(".", "") for x in [str(x) for x in list_to_search]]
             )